+++ pkg/client/connector/install_actions.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"path/filepath"
 	"strconv"
 	"strings"
 
@@ -148,26 +147,6 @@ func marshalString(data completeAction) (string, error) {
 
 func unmarshalString(in string, out completeAction) error {
 	return json.Unmarshal([]byte(in), out)
-}
-
-func GetPodTemplateFromObject(obj kates.Object) (*kates.PodTemplateSpec, error) {
-	var tplSpec *kates.PodTemplateSpec
-	kind := obj.GetObjectKind().GroupVersionKind().Kind
-	switch kind {
-	case "ReplicaSet":
-		rs := obj.(*kates.ReplicaSet)
-		tplSpec = &rs.Spec.Template
-	case "Deployment":
-		dep := obj.(*kates.Deployment)
-		tplSpec = &dep.Spec.Template
-	case "StatefulSet":
-		statefulSet := obj.(*kates.StatefulSet)
-		tplSpec = &statefulSet.Spec.Template
-	default:
-		return nil, objErrorf(obj, "unsupported workload kind %q", kind)
-	}
-
-	return tplSpec, nil
 }
 
 // A makePortSymbolicAction replaces the numeric TargetPort of a ServicePort with a generated
@@ -341,11 +320,6 @@ func (s *svcActions) TelVersion() (semver.Version, error) {
 
 // addTrafficAgentAction ///////////////////////////////////////////////////////
 
-const (
-	envPrefix        = "TEL_APP_"
-	telAppMountPoint = "/tel_app_mounts"
-)
-
 // addTrafficAgentAction is a partialAction that adds a traffic-agent to the set of containers in a
 // pod template spec.
 type addTrafficAgentAction struct {
@@ -384,137 +358,10 @@ func (ata *addTrafficAgentAction) Do(obj kates.Object) error {
 		return objErrorf(obj, "unable to find app container %q in", ata.containerName)
 	}
 
-	tplSpec.Spec.Volumes = append(tplSpec.Spec.Volumes, corev1.Volume{
-		Name: agentAnnotationVolumeName,
-		VolumeSource: corev1.VolumeSource{
-			DownwardAPI: &corev1.DownwardAPIVolumeSource{
-				Items: []corev1.DownwardAPIVolumeFile{
-					{
-						FieldRef: &corev1.ObjectFieldSelector{
-							FieldPath: "metadata.annotations",
-						},
-						Path: "annotations",
-					},
-				},
-			},
-		},
-	})
-	tplSpec.Spec.Containers = append(tplSpec.Spec.Containers, corev1.Container{
-		Name:  agentContainerName,
-		Image: ata.ImageName,
-		Args:  []string{"agent"},
-		Ports: []corev1.ContainerPort{{
-			Name:          ata.ContainerPortName,
-			Protocol:      ata.ContainerPortProto,
-			ContainerPort: 9900,
-		}},
-		Env:          ata.agentEnvironment(obj.GetName(), appContainer),
-		EnvFrom:      ata.agentEnvFrom(appContainer.EnvFrom),
-		VolumeMounts: ata.agentVolumeMounts(appContainer.VolumeMounts),
-		ReadinessProbe: &corev1.Probe{
-			Handler: corev1.Handler{
-				Exec: &corev1.ExecAction{
-					Command: []string{"/bin/stat", "/tmp/agent/ready"},
-				},
-			},
-		},
-	})
+	tplSpec.Spec.Volumes = append(tplSpec.Spec.Volumes, AgentVolume())
+	tplSpec.Spec.Containers = append(tplSpec.Spec.Containers,
+		AgentContainer(ata, obj, appContainer))
 	return nil
-}
-
-func (ata *addTrafficAgentAction) agentEnvFrom(appEF []corev1.EnvFromSource) []corev1.EnvFromSource {
-	if ln := len(appEF); ln > 0 {
-		agentEF := make([]corev1.EnvFromSource, ln)
-		for i, appE := range appEF {
-			appE.Prefix = envPrefix + appE.Prefix
-			agentEF[i] = appE
-		}
-		return agentEF
-	}
-	return appEF
-}
-
-func (ata *addTrafficAgentAction) agentEnvironment(agentName string, appContainer *kates.Container) []corev1.EnvVar {
-	appEnv := ata.appEnvironment(appContainer)
-	env := make([]corev1.EnvVar, len(appEnv), len(appEnv)+7)
-	copy(env, appEnv)
-	env = append(env,
-		corev1.EnvVar{
-			Name:  "LOG_LEVEL",
-			Value: "debug",
-		},
-		corev1.EnvVar{
-			Name:  "AGENT_NAME",
-			Value: agentName,
-		},
-		corev1.EnvVar{
-			Name: "AGENT_NAMESPACE",
-			ValueFrom: &corev1.EnvVarSource{
-				FieldRef: &corev1.ObjectFieldSelector{
-					FieldPath: "metadata.namespace",
-				},
-			},
-		},
-		corev1.EnvVar{
-			Name: "AGENT_POD_IP",
-			ValueFrom: &corev1.EnvVarSource{
-				FieldRef: &corev1.ObjectFieldSelector{
-					FieldPath: "status.podIP",
-				},
-			},
-		},
-		corev1.EnvVar{
-			Name:  "APP_PORT",
-			Value: strconv.Itoa(int(ata.ContainerPortNumber)),
-		})
-	if len(appContainer.VolumeMounts) > 0 {
-		env = append(env, corev1.EnvVar{
-			Name:  "APP_MOUNTS",
-			Value: telAppMountPoint,
-		})
-
-		// Have the agent propagate the mount-points as TELEPRESENCE_MOUNTS to make it easy for the
-		// local app to create symlinks.
-		mounts := make([]string, len(appContainer.VolumeMounts))
-		for i := range appContainer.VolumeMounts {
-			mounts[i] = appContainer.VolumeMounts[i].MountPath
-		}
-		env = append(env, corev1.EnvVar{
-			Name:  envPrefix + "TELEPRESENCE_MOUNTS",
-			Value: strings.Join(mounts, ":"),
-		})
-	}
-	env = append(env, corev1.EnvVar{
-		Name:  "MANAGER_HOST",
-		Value: managerAppName + "." + managerNamespace,
-	})
-	return env
-}
-
-func (ata *addTrafficAgentAction) agentVolumeMounts(mounts []corev1.VolumeMount) []corev1.VolumeMount {
-	agentMounts := make([]corev1.VolumeMount, len(mounts)+1)
-	for i, mount := range mounts {
-		mount.MountPath = filepath.Join(telAppMountPoint, mount.MountPath)
-		agentMounts[i] = mount
-	}
-	agentMounts[len(mounts)] = corev1.VolumeMount{
-		Name:      agentAnnotationVolumeName,
-		MountPath: "/tel_pod_info",
-	}
-	return agentMounts
-}
-
-func (ata *addTrafficAgentAction) appEnvironment(appContainer *kates.Container) []corev1.EnvVar {
-	envCopy := make([]corev1.EnvVar, len(appContainer.Env)+1)
-	for i, ev := range appContainer.Env {
-		ev.Name = envPrefix + ev.Name
-		envCopy[i] = ev
-	}
-	envCopy[len(appContainer.Env)] = corev1.EnvVar{
-		Name:  "TELEPRESENCE_CONTAINER",
-		Value: appContainer.Name,
-	}
-	return envCopy
 }
 
 func (ata *addTrafficAgentAction) ExplainDo(_ kates.Object, out io.Writer) {
