+++ pkg/client/connector/installutil_constants.go
@@ -1,0 +1,10 @@
+package connector
+
+const (
+	agentContainerName        = "traffic-agent"
+	agentAnnotationVolumeName = "traffic-annotations"
+	domainPrefix              = "telepresence.getambassador.io/"
+	managerAppName            = "traffic-manager"
+	ManagerPortHTTP           = 8081
+	telAppMountPoint          = "/tel_app_mounts"
+)
